+++ src/lib.rs
@@ -17,6 +17,7 @@
 
 #![allow(dead_code)]
 #![allow(non_camel_case_types)]
+#![forbid(unsafe_code)]
 
 mod parquet_format;
 pub use crate::parquet_format::*;
